+++ agent/proxyprocess/proxy_test.go
@@ -1,4 +1,4 @@
-package proxy
+package proxyprocess
 
 import (
 	"fmt"
