+++ agent/proxyprocess/root.go
@@ -1,4 +1,4 @@
-package proxy
+package proxyprocess
 
 import (
 	"os"
